feat: add stdin-driven simulator control for agent debugging (button inject, screenshots, BMP→PNG converter), wire screenshot capture into the main loop, document the visual-debug workflow, fix ADVERT lat/lon skip to 8 bytes, cast T4 counter to unsigned, and drop dead hub contact-loading locals

+++ src/activities/util/T4EntryActivity.cpp
@@ -1175,7 +1175,7 @@ int T4EntryActivity::renderCandidateRow(int startY) {
 
   // ── Counter: "15 / 228" centered below ──────────────────────────────
   char counter[16];
-  snprintf(counter, sizeof(counter), "%u / %u", activeIdx + 1, candCount);
+  snprintf(counter, sizeof(counter), "%u / %u", static_cast<unsigned>(activeIdx + 1), static_cast<unsigned>(candCount));
   int counterW = renderer.getTextWidth(UI_12_FONT_ID, counter);
   int counterX = effectiveMargin + (textAreaWidth - counterW) / 2;
   int counterY = startY + renderer.getLineHeight(UI_12_FONT_ID) + 4;
